OnMarkOverEvent 變更為獨立 class
