+++ src/main/java/us/dontcareabout/npm/client/ui/UiCenter.java
@@ -5,14 +5,14 @@ import com.google.gwt.event.shared.HandlerRegistration;
 import com.google.gwt.event.shared.SimpleEventBus;
 
 public class UiCenter {
-	public final static SimpleEventBus eventBus = new SimpleEventBus();
+	private final static SimpleEventBus eventBus = new SimpleEventBus();
 
 	public static void fire(GwtEvent<?> event) {
 		eventBus.fireEvent(event);
 	}
 
-	public static HandlerRegistration addMarkOverHandler(TimelineLayer.OnMarkOverHandler handler) {
-		return eventBus.addHandler(TimelineLayer.OnMarkOverEvent.TYPE, handler);
+	public static HandlerRegistration addMarkOverHandler(OnMarkOverEvent.OnMarkOverHandler handler) {
+		return eventBus.addHandler(OnMarkOverEvent.TYPE, handler);
 	}
 
 	public static HandlerRegistration addMarkLeaveHandler(TimelineLayer.OnMarkLeaveHandler handler) {
